Module laminarmq::commit_log

Module providing a “commit log” abstraction.

This kind of data structure is primarily useful for storing a series of “events”. This module provides the abstractions necessary for manipulating this data structure asynchronously on persistent storage media. The abstractions provided by this module are generic over different async runtimes and storage media. Hence user’s will need to choose their specializations and implementations of these generic abstractions as necessary.

Also see: segmented_log, glommio_impl.

In the context of laminarmq this module is intended to provide the storage for individual partitions in a topic.

Modules

glommio_impl

Module providing specializations of different generic entities and trait impls for the [glommio] runtime.

prelude

Prelude module for commit_log with common exports for convenience.

segmented_log

Module providing the SegmentedLog abstraction for supporting a immutable, segmented and persistent CommitLog implementation.

Structs

Record

Represents a record in a CommitLog.

Traits

CommitLog

An ordered sequential collection of Record instances.

Functions

commit_log_record_stream

Returns a [Stream] of Record instances from the records stored in this CommitLog implementation instance. This function reads records starting from the given offset linearly upto the CommitLog::highest_offset of this commit log.