Function laminarmq::commit_log::commit_log_record_stream

pub fn commit_log_record_stream<'log, M, T, CL: CommitLog<M, T>>(
    commit_log: &'log CL,
    from_offset: u64,
    scan_seek_bytes: u64
) -> impl Stream<Item = Record<M, T>> + 'logwhere
    M: Serialize + DeserializeOwned + 'log,
    T: Deref<Target = [u8]> + 'log,

Returns a [Stream] of Record instances from the records stored in this CommitLog implementation instance. This function reads records starting from the given offset linearly upto the CommitLog::highest_offset of this commit log.

Params:

commit_log: The CommitLog implementation instance to read records from. from_offset: The offset from which to start reading records. scan_seek_bytes: The number of bytes to seek forward and scan for valid records when an error occurs.

Returns:

A [Stream] containing the Records in this commit log in FIFO order.