Struct laminarmq::commit_log::glommio_impl::segmented_log::store::Store

pub struct Store { /* private fields */ }

crate::commit_log::segmented_log::store::Store implementation for the [glommio] runtime.

This implementation uses directly mapped files to leverage io_uring powered operations. A Store consists of a [DmaFile] reader and [DmaStreamWriter], both of which point to the same underlying file on the disk.

We have chosen a simple [DmaFile] for the reader since the read amplification caused due to random unbuffered reads is manageable. However, write amplifications are generally more costly. Since we always append to the end of the file, [DmaStreamWriter] is a better solution which enables use to have the best tradeoff of responsiveness v/s buffering.

Implementations

impl Store

pub async fn new<P: AsRef<Path>>(path: P) -> Result<Self, StoreError>

Creates a new Store from a [glommio::io::DmaFile]

pub async fn with_path_and_buffer_size<P: AsRef<Path>>(
path: P,
buffer_size: usize
) -> Result<Self, StoreError>

Creates a new Store instance.

The writer is opened with “write”, “append” and “create” flags, while the reader is opened with only the “read” flag. The given buffer_size is used to configure the buffer size of the [DmaStreamWriter].