Module laminarmq::commit_log::segmented_log::common

Module providing utilities used by SegmentedLog.

Functions

obtain_store_files_in_directory

Returns an iterator of the paths of all the super::store::Store backing files in the given directory.

read_and_write_segments

Seperates the read segments and write segments in the given Vec<Segment>.

segments_in_directory

Returns a Vec<Segment> of all the segments stored in this directory.

store_file_path

Returns the backing Store file path, with the given Log storage directory and the given segment base offset.

store_paths_sorted_by_offset

Returns the given store paths sorted by the base offset of the respective segments that they belong to.