Function laminarmq::commit_log::segmented_log::common::read_and_write_segments

pub async fn read_and_write_segments<T, M, S, SegC, P: AsRef<Path>>(
    storage_dir_path: P,
    segment_creator: &SegC,
    log_config: SegmentedLogConfig,
    segments: Vec<Segment<T, M, S>>
) -> Result<(Vec<Segment<T, M, S>>, Segment<T, M, S>), SegmentedLogError<T, S>>where
    T: Deref<Target = [u8]> + SplitAt<u8>,
    M: Default + Serialize + DeserializeOwned,
    S: Store<T>,
    SegC: SegmentCreator<T, M, S>,

Seperates the read segments and write segments in the given Vec<Segment>.

There are the following cases that can happen here:

If the given vector of segments is empty, we return an empty vector of read segments and a new write segment.

If the last segment is the given list of segment is maxed out, we return the given vector of segments as is as the vector of read segments, along with a new write segment.

Otherwise we pop out the last segment from the given vector of segments. The shorter vector of segments is returned as the vector of read segments along with the popped out segment as the write segment.

Errors

SegmentedLogError::SegmentError: if there was an error in creating a new segment. (When creating the write segment)

SegmentedLogError::WriteSegmentLost: if the popped last segments from the given list of segments is None when segments.last() returned a [Some(_)].