Function laminarmq::commit_log::segmented_log::common::store_paths_sorted_by_offset

pub fn store_paths_sorted_by_offset(
    store_paths: impl Iterator<Item = PathBuf>
) -> Vec<(PathBuf, u64)>

Returns the given store paths sorted by the base offset of the respective segments that they belong to.

Returns

A vector of (path, offset) tuples, sorted by the offsets.