Module laminarmq::commit_log::segmented_log::segment

Module providing cross platform Segment abstractions for reading and writing Record instances from Store instances.

Modules

config

Module providing types for configuring Segment instances.

Structs

Segment

Cross platform storage abstraction used for reading and writing Record instances from Store implementations. It also acts the unit of storage in a SegmentedLog.

Enums

SegmentError

Error type used for operations on a Segment.

Functions

segment_record_stream

Returns a [Stream] of Records in the given Segment starting from the given offset in FIFO order.