Struct laminarmq::commit_log::segmented_log::segment::Segment

pub struct Segment<T, M, S: Store<T>>where
    T: Deref<Target = [u8]> + SplitAt<u8>,
    M: Default + Serialize + DeserializeOwned,{ /* private fields */ }

Cross platform storage abstraction used for reading and writing Record instances from Store implementations. It also acts the unit of storage in a SegmentedLog.

Segment uses binary encoding with bincode for serializing Records.

Each segment can store a fixed size of bytes. A segment has a base_offset which is the offset of the first message in the segment. Note that the base_offset is not necessarily zero and can be arbitrarily high number. The next_offset for a segment is offset at which the next record will be written. (offset of last record + size of last record.)

[segment]
┌────────────────────────────────────────────────────────────────┐
│                                                                │
│  record offset = position in store + segment base_offset       │
│                                                                │
│  [store]                                                       │
│  ┌──────────┐                                                  │
│  │ record#x │ position: 0                                      │
│  ├──────────┤                                                  │
│  │ record#y │ position: size(record#x)                         │
│  ├──────────┤                                                  │
│  │ record#z │ position: size(record#x) + size(record#y)        │
│  ├──────────┤                                                  │
│      ....                                                      │
│                                                                │
└────────────────────────────────────────────────────────────────┘

Each segment is backed by a store which is a handle to a file on disk. Since it directly deals with file, position starts from 0, like file position. As a consequence, we translate store position to segment offset by simply adding the segment offset to the store position. Writes in a store are buffered. The write buffer is synced when it fills up, to persist data to the disk. Reads read values that have been successfully synced and persisted to the disk. Hence, reads on a store lag behind the writes.

All writes go to the write segment. When we max out the capacity of the write segment, we close the write segment and reopen it as a read segment. The re-opened segment is added to the list of read segments. A new write segment is then created with base_offset as the next_offset of the previous write segment.

Implementations

impl<T, M, S: Store<T>> Segment<T, M, S>where
T: Deref<Target = [u8]> + SplitAt<u8>,
M: Default + Serialize + DeserializeOwned,

pub fn with_config_base_offset_and_store(
config: SegmentConfig,
base_offset: u64,
store: S
) -> Self

Creates a segment instance with the given config, base offset and store instance.

pub fn base_offset(&self) -> u64

Offset of the first record in this segment instance.

pub fn next_offset(&self) -> u64

Offset at which the next record will be written in this segment.

pub fn store(&self) -> &S

Returns a reference to the backing Store.

pub fn store_position(&self, offset: u64) -> Option<u64>

Returns the position on the underlying store to which the given offset would map to. This method returns None if the given offset is beyond the capacity.

pub fn is_maxed(&self) -> bool

Returns whether the size of the underlying store in number of bytes is greater than or equal to the allowed limit in config::SegmentConfig::max_store_bytes.

pub fn size(&self) -> u64

Returns the size of the underlying store.

pub async fn append(
&mut self,
record_bytes: &[u8],
metadata: RecordMetadata<M>
) -> Result<(u64, usize), SegmentError<T, S>>

Appends the given record to the end of this store.

Returns

The offset at which the record was written, along with the number of bytes written.

Errors

SegmentError::SegmentMaxed if Self::is_maxed returns true.

SegmentError::SerializationError if there was an error during binary encoding the given record instance.

SegmentError::StoreError if there was an error during writing to the underlying Store instance.

pub fn offset_within_bounds(&self, offset: u64) -> bool

pub async fn read(
&self,
offset: u64
) -> Result<(Record<RecordMetadata<M>, T>, u64), SegmentError<T, S>>

Reads the record at the given offset.

Returns

A tuple containing:

Record instance containing the desired record.

u64 value containing the offset of the next record.

Errors

SegmentError::OffsetOutOfBounds if the offset >= next_offset value.

SegmentError::OffsetBeyondCapacity if the given offset is beyond this stores capacity, and doesn’t map to a valid position on the underlying Store instance.

SegmentError::StoreError if there was an error during reading the record at the given offset from the underlying Store instance.

SegmentError::SerializationError if there is an error during deserializing the record from the bytes read from storage.

pub async fn truncate(&mut self, offset: u64) -> Result<(), SegmentError<T, S>>

pub async fn remove(self) -> Result<(), SegmentError<T, S>>

Removes the underlying Store instances associated. This method consumes this Segment instance to prevent further operations on this segment.

Errors

SegmentError::StoreError if there was an error in removing the underlying store.

pub async fn close(self) -> Result<(), SegmentError<T, S>>

Closes the underlying Store instances associated. This method consumes this Segment instance to prevent further operations on this segment.

Errors

SegmentError::StoreError if there was an error in closing the underlying store.

pub fn creation_time(&self) -> Instant

Returns an std::time::Instant value denoting the time at which this segment was created.