Trait laminarmq::commit_log::segmented_log::store::Store

pub trait Store<T>where
    T: Deref<Target = [u8]>,{
    type Error: Error;

    fn append_multipart<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        record_bytes: &'life1 [&'life2 [u8]]
    ) -> Pin<Box<dyn Future<Output = Result<(u64, usize), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn read<'life0, 'async_trait>(
        &'life0 self,
        position: u64
    ) -> Pin<Box<dyn Future<Output = Result<(T, u64), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn truncate<'life0, 'async_trait>(
        &'life0 mut self,
        position: u64
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn close<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        Self: 'async_trait;
    fn remove<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        Self: 'async_trait;
    fn size(&self) -> u64;
    fn path(&self) -> Result<Ref<'_, Path>, Self::Error>;

    fn append<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        record_bytes: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<(u64, usize), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Trait representing a collection of record backed by some form of persistent storage.

This trait is generic over the kind of records that can be stored, provided that the can be dereferenced as bytes. In the context of a segmented log, they act as the backing storage for a “segment”.

This trait’s contract assumes the following with regard to how records are laid out in storage:

┌─────────────────────────┬────────────────────────┬───────────────────────┐
│ crc32_checksum: [u8; 4] │ record_length: [u8; 4] │ record_bytes: [u8; _] │
└─────────────────────────┴────────────────────────┴───────────────────────┘

As such if the record header is invalidated due to checksum mismatch or record length mismatch an appropriate error is expected on the associated operation.

Required Associated Types

type Error: Error

The error type used by the methods of this trait.

Required Methods

fn append_multipart<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
record_bytes: &'life1 [&'life2 [u8]]
) -> Pin<Box<dyn Future<Output = Result<(u64, usize), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Appends a record split in multiple parts as a single record at the end of this store.

fn read<'life0, 'async_trait>(
&'life0 self,
position: u64
) -> Pin<Box<dyn Future<Output = Result<(T, u64), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the record stored at the given position in the Store along with the position of the next record.

fn truncate<'life0, 'async_trait>(
&'life0 mut self,
position: u64
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Truncates this Store instance by removing all records after the given position and makes the given position the size of this store. Returns a suitable error if this position is not a valid record position.

fn close<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
Self: 'async_trait,

Closes this record. Consumes this record instance to stop further operations on this closed store instance.

fn remove<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
Self: 'async_trait,

Removes the file associated with this store from persistent media. Consumes this record to stop further operations on this “removed” store instance.

fn size(&self) -> u64

Number of bytes stored in/with this store instance.

fn path(&self) -> Result<Ref<'_, Path>, Self::Error>

Returns the path to the underlying file used for storage.

Provided Methods

fn append<'life0, 'life1, 'async_trait>(
&'life0 mut self,
record_bytes: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<(u64, usize), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Appends a record containing the given record bytes at the end of this store.