Trait laminarmq::commit_log::segmented_log::SegmentCreator

pub trait SegmentCreator<T, M, S>where
    T: Deref<Target = [u8]> + SplitAt<u8>,
    M: Default + Serialize + DeserializeOwned,
    S: Store<T>,{
    fn new_segment_with_store_file_path_offset_and_config<'life0, 'async_trait, P>(
        &'life0 self,
        store_file_path: P,
        base_offset: u64,
        config: SegmentConfig
    ) -> Pin<Box<dyn Future<Output = Result<Segment<T, M, S>, SegmentError<T, S>>> + 'async_trait>>
    where
        P: 'async_trait + AsRef<Path>,
        'life0: 'async_trait,
        Self: 'async_trait;

    fn new_segment_with_storage_dir_offset_and_config<'life0, 'async_trait, P>(
        &'life0 self,
        storage_dir: P,
        base_offset: u64,
        config: SegmentConfig
    ) -> Pin<Box<dyn Future<Output = Result<Segment<T, M, S>, SegmentError<T, S>>> + 'async_trait>>
    where
        P: 'async_trait + AsRef<Path>,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Strategy pattern trait for constructing Segment instances for SegmentedLog.

Required Methods

fn new_segment_with_store_file_path_offset_and_config<'life0, 'async_trait, P>(
&'life0 self,
store_file_path: P,
base_offset: u64,
config: SegmentConfig
) -> Pin<Box<dyn Future<Output = Result<Segment<T, M, S>, SegmentError<T, S>>> + 'async_trait>>where
P: 'async_trait + AsRef<Path>,
'life0: 'async_trait,
Self: 'async_trait,

Constructs a new segment with it’s store at the given store path, base_offset and segment config.

Provided Methods

fn new_segment_with_storage_dir_offset_and_config<'life0, 'async_trait, P>(
&'life0 self,
storage_dir: P,
base_offset: u64,
config: SegmentConfig
) -> Pin<Box<dyn Future<Output = Result<Segment<T, M, S>, SegmentError<T, S>>> + 'async_trait>>where
P: 'async_trait + AsRef<Path>,
'life0: 'async_trait,
Self: 'async_trait,

Constructs a new segment in the given storage directory with the given base offset and config.

The default provided implementation simply invokes Self::new_segment_with_store_file_path_offset_and_config with store_file_path obtained by invoking common::store_file_path on storage_dir and base_offset.