Trait laminarmq::commit_log::CommitLog

pub trait CommitLog<M, T>where
    M: Serialize + DeserializeOwned,
    T: Deref<Target = [u8]>,{
    type Error;

    fn highest_offset(&self) -> u64;
    fn lowest_offset(&self) -> u64;
    fn append<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        record_bytes: &'life1 [u8],
        metadata: M
    ) -> Pin<Box<dyn Future<Output = Result<(u64, usize), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn read<'life0, 'async_trait>(
        &'life0 self,
        offset: u64
    ) -> Pin<Box<dyn Future<Output = Result<(Record<M, T>, u64), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn truncate<'life0, 'async_trait>(
        &'life0 mut self,
        _offset: u64
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn remove<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        Self: 'async_trait;
    fn close<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        Self: 'async_trait;

    fn offset_within_bounds(&self, offset: u64) -> bool { ... }
    fn remove_expired<'life0, 'async_trait>(
        &'life0 mut self,
        _expiry_duration: Duration
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
}

An ordered sequential collection of Record instances.

A Record in an CommitLog is addressed with an unique u64 offset, which denotes it position in the collection of records. The offsets in a CommitLog are monotonically increasing. A higher offset denotes that the Record appears later in the CommitLog.

This is useful for storing a sequence of events or operations, which may be used to restore a system to it’s previous or current state.

Required Associated Types

type Error

Error type used by the methods of this trait.

Required Methods

fn highest_offset(&self) -> u64

Returns the offset where the next Record will be written in this CommitLog.

fn lowest_offset(&self) -> u64

Returns the offset of the first Record in this CommitLog.

fn append<'life0, 'life1, 'async_trait>(
&'life0 mut self,
record_bytes: &'life1 [u8],
metadata: M
) -> Pin<Box<dyn Future<Output = Result<(u64, usize), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Appends a new Record, containing the given record bytes and with the given metadata at the end of this CommitLog.

Returns:

The offset at which the record was written, along with the number of bytes written.

Errors

Possible errors arising in implementations could include:

Invalid record metadata

Errors from underlying storage media

fn read<'life0, 'async_trait>(
&'life0 self,
offset: u64
) -> Pin<Box<dyn Future<Output = Result<(Record<M, T>, u64), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Reads the Record at the given offset, along with the offset of the next record from this CommitLog.

Returns:

A tuple containing the Record instance along with the u64 offset.

Errors:

Possible errors arising in implementations could include:

Invalid read offset

Errors from underlying storage media

fn truncate<'life0, 'async_trait>(
&'life0 mut self,
_offset: u64
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Truncates this CommitLog instance by removing all records starting from the given offset. The offset is expected to be at the start or end of a record.

Errors:

Possible errors arising in implementations could include:

Invalid truncateion offset (offset out of bounds or in the middle of a record)

Errors from underlying storage media during remooval of backing files.

fn remove<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
Self: 'async_trait,

Removes all underlying storage files associated. Consumes this CommitLog instance to prevent further operations on this instance.

Errors:

Possible errors arising in implementations could include:

Errors from underlying storage media during removal of backing files.

fn close<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
Self: 'async_trait,

Closes the files associated with this CommitLog instances and syncs them to storage media. Consumes this CommitLog instance to prevent further operations on this instance.

Errors:

Possible errors arising in implementations could include:

Errors from underlying storage media when syncing and closing files.

Provided Methods

fn offset_within_bounds(&self, offset: u64) -> bool

Returns whether the given offset lies in [lowest_offset, highest_offset)

fn remove_expired<'life0, 'async_trait>(
&'life0 mut self,
_expiry_duration: Duration
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Remove expired storage used, if any. Default implementation simply returns with [Ok(())]

Errors:

Possible errors arising in implementations could include errors during removal of files.