Module laminarmq::common::cache

Structs

AllocBTreeMap

A Map implementation based on alloc::collections::BTreeMap.

AllocVec

Implements [Vector] with alloc::vec::Vec.

LRUCache

A generational-arena powered LRUCache implementation.

Link

Represents a link to node in the linked list.

NoOpCache

UnsupportedOp

Enums

Eviction

An evicted value from cache.

LRUCacheError

Lookup

Traits

Cache

A size bounded map, where certain existing entries are evicted to make space for new entires.

Map

An abstrct mapping from a set of keys to a set of values.

Type Aliases

AllocLRUCache

LRUCacheBlockArenaEntry

Alias representing block entries for storage in a generational arena.