Trait laminarmq::common::cache::Cache

pub trait Cache<K, V> {
    type Error;

    // Required methods
    fn insert(
        &mut self,
        key: K,
        value: V
    ) -> Result<Eviction<K, V>, Self::Error>;
    fn remove(&mut self, key: &K) -> Result<Lookup<V>, Self::Error>;
    fn shrink(&mut self, new_capacity: usize) -> Result<(), Self::Error>;
    fn reserve(&mut self, additional: usize) -> Result<(), Self::Error>;
    fn query(&mut self, key: &K) -> Result<Lookup<&V>, Self::Error>;
    fn capacity(&self) -> usize;
    fn len(&self) -> usize;
    fn clear(&mut self) -> Result<(), Self::Error>;

    // Provided methods
    fn is_maxed(&self) -> bool { ... }
    fn is_empty(&self) -> bool { ... }
}

A size bounded map, where certain existing entries are evicted to make space for new entries.

Implementations follow a well defined criteria to decide which cache blocks to evict in which order. (e.g an LRU cache implementation would evict the least recently used cache blocks).

Required Associated Types

type Error

Associated error type.

Required Methods

fn insert(&mut self, key: K, value: V) -> Result<Eviction<K, V>, Self::Error>

Inserts the given key/value pair into this cache.

fn remove(&mut self, key: &K) -> Result<Lookup<V>, Self::Error>

Removes the key/value pair associated with the given key from this cache.

fn shrink(&mut self, new_capacity: usize) -> Result<(), Self::Error>

Removes (self.len() - new_capacity) cache blocks to fit the new capacity. If the difference is non-positive no cache blocks are removed.

fn reserve(&mut self, additional: usize) -> Result<(), Self::Error>

Reserves additional memory to accomodate the given number of additional cache blocks.

fn query(&mut self, key: &K) -> Result<Lookup<&V>, Self::Error>

Queries this cache to find the value associated with given key.

fn capacity(&self) -> usize

Returns the current capacity of this cache.

fn len(&self) -> usize

Returns the number of key/value pairs stored in this cache.

fn clear(&mut self) -> Result<(), Self::Error>

Remove all items from this cache until it’s empty.

Provided Methods

fn is_maxed(&self) -> bool

Returns whether this cache is maxed out.

This method simply checks whether the current length of this cache equals its capacity.

fn is_empty(&self) -> bool

Returns whether this cache is empty.