Type Alias laminarmq::common::cache::LRUCacheBlockArenaEntry

pub type LRUCacheBlockArenaEntry<K, T> = Entry<Node<Block<K, T>>>;

Alias representing block entries for storage in a generational arena.

Aliased Type

enum LRUCacheBlockArenaEntry<K, T> {
    Occupied {
        value: Node<Block<K, T>>,
        generation: u64,
    },
    Free {
        next_free_idx: Option<usize>,
    },
    Unmapped,
}

Variants

Occupied

Fields

value: Node<Block<K, T>>

generation: u64

An occupied entry containing an allocated value and the associated generation counter.

Free

Fields

next_free_idx: Option<usize>

Free entry pointing to next free entry in the free list.

Unmapped

An unmapped arena entry.