Trait laminarmq::common::serde_compat::SerializationProvider

pub trait SerializationProvider {
    type SerializedBytes: Deref<Target = [u8]>;
    type Error: Error;

    // Required methods
    fn serialize<T>(value: &T) -> Result<Self::SerializedBytes, Self::Error>
       where T: Serialize;
    fn serialized_size<T>(value: &T) -> Result<usize, Self::Error>
       where T: Serialize;
    fn deserialize<'a, T>(bytes: &'a [u8]) -> Result<T, Self::Error>
       where T: Deserialize<'a>;
}

Trait to represent a serialization provider.

Required Associated Types

type SerializedBytes: Deref<Target = [u8]>

Serialized bytes container.

type Error: Error

Error type used by the fallible functions of this trait.

Required Methods

fn serialize<T>(value: &T) -> Result<Self::SerializedBytes, Self::Error>
where T: Serialize,

Serializes the given value.

fn serialized_size<T>(value: &T) -> Result<usize, Self::Error>
where T: Serialize,

Returns the number of bytes used by the serialized representation of value.

fn deserialize<'a, T>(bytes: &'a [u8]) -> Result<T, Self::Error>
where T: Deserialize<'a>,

Deserializes the given serialized bytes into a T instance.

Object Safety

This trait is not object safe.