Trait laminarmq::common::split::SplitAt

pub trait SplitAt<T>: Deref<Target = [T]> + Sized {
    fn split_at(self, at: usize) -> Option<(Self, Self)>;
}

Trait respresenting a slice that can be split at a given position.

Required Methods

fn split_at(self, at: usize) -> Option<(Self, Self)>

Splits this slice at the given position. The left half contains elements in the [0, at) index range while the right half contains elements in the [at, self.len()) index range.