Trait laminarmq::server::channel::Sender

pub trait Sender<T> {
    type Error: Error;

    // Required methods
    fn send<'life0, 'async_trait>(
        &'life0 self,
        item: T
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn try_send(&self, item: T) -> Result<(), Self::Error>;
}

Trait representing the sending end of a channel.

Required Associated Types

type Error: Error

Required Methods

fn send<'life0, 'async_trait>( &'life0 self, item: T ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Sends the given valye of this channel. Asynchronously waits if there is no space to write an item to the channel.

Errors

Possible error situations could include:

unable to send item

receiving end dropped

fn try_send(&self, item: T) -> Result<(), Self::Error>

Sends the given value over this channel. This method is expected not to block and return immediately.

Errors

Possible error situations could include:

unable to send item

receiving end dropped