Function laminarmq::server::glommio_impl::worker::new_task

pub fn new_task<P: Partition, Request, Response>(
    request: Request
) -> (GlommioTask<P, Request, Response>, ResponseReceiver<Response, P>)

Creates a new GlommioTask for servicing the given Request.

Returns:

GlommioTask: Task to be executed by processor.

ResponseReceiver: receiving end of the response channel where the response will be received after the request is processed.