Trait laminarmq::server::partition::Partition

pub trait Partition {
    type Error: Error;
    type Request;
    type Response;

    fn serve_idempotent<'life0, 'async_trait>(
        &'life0 self,
        request: Self::Request
    ) -> Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn serve<'life0, 'async_trait>(
        &'life0 mut self,
        request: Self::Request
    ) -> Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn close<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        Self: 'async_trait;
    fn remove<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>
    where
        Self: 'async_trait;
}

Unit of data sharding and replication. A partition provides an abstract RPC service for serving record read/write requests.

Required Associated Types

type Error: Error

type Request

type Response

Required Methods

fn serve_idempotent<'life0, 'async_trait>(
&'life0 self,
request: Self::Request
) -> Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Serves idempotent requests from a shared partition ref that have no side effects.

fn serve<'life0, 'async_trait>(
&'life0 mut self,
request: Self::Request
) -> Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Exclusively serves requests from a mutable partition ref, that may or may not have side effects.

fn close<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
Self: 'async_trait,

Closes all files associated with this partition and syncs all changes to disk. This method acquires the ownership of this partition to prevent this partition from being used any further.

fn remove<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + 'async_trait>>where
Self: 'async_trait,

Removes all storage files associated with this partition. This method acquires the ownership of this partition to prevent this partition from being used any further.