Module laminarmq::server::worker::single_node

Module providing single node specific request and response types.

Enums

ProcessorRequest

Adminstrative requests specific to a Processor and not a specific Partition instance.

WorkerRequest

Request enumeration to generalize over ProcessorRequest and PartitionRequest.