Module laminarmq::storage::commit_log::segmented_log::index

Provides components necessary for mapping record indices to store-file positions in segments.

This module provides Index which maintains a mapping from logical record indices to positions on the Store of a Segment.

Structs

Index

Index for every Segment in a SegmentedLog.

IndexBaseMarker

Marker to persist the starting base_index of an Index in the index Storage.

IndexRecord

Unit of storage on an Index. Stores position, length and checksum metadata for a single Record persisted in the segment Store.

Enums

IndexError

Error type associated with operations on Index.

Constants

INDEX_BASE_MARKER_LENGTH

Number of bytes required for storing the base marker.

INDEX_BASE_POSITION

Lowest underlying storage position

INDEX_FILE_EXTENSION

Extension used by backing files for Index instances.

INDEX_RECORD_LENGTH

Number of bytes required for storing the record header.