Trait laminarmq::storage::commit_log::segmented_log::segment::SegmentStorageProvider

pub trait SegmentStorageProvider<S, Idx>
where
    S: Storage,
{
    // Required methods
    fn obtain_base_indices_of_stored_segments<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Idx>, S::Error>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn obtain<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        idx: &'life1 Idx
    ) -> Pin<Box<dyn Future<Output = Result<SegmentStorage<S>, S::Error>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Provides backing storage for Segment instances.

Used to abstract the mechanism of acquiring storage handles from the underlying persistent media.

Required Methods

fn obtain_base_indices_of_stored_segments<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<Vec<Idx>, S::Error>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the base indices of all the Segment instances persisted in this storage media.

fn obtain<'life0, 'life1, 'async_trait>( &'life0 mut self, idx: &'life1 Idx ) -> Pin<Box<dyn Future<Output = Result<SegmentStorage<S>, S::Error>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Obtains a SegmentStorage instance for a Segment with the given idx as their base index.

Implementations are required to allocate/arrange new storage handles if a Segment with the given base index is not already persisted on the underlying storage media.