Type Alias laminarmq::storage::commit_log::segmented_log::LogError

pub type LogError<S, SERP, C> = SegmentedLogError<<S as Storage>::Error, <SERP as SerializationProvider>::Error, <C as Cache<usize, ()>>::Error>;

Type alias for SegmentedLogError with additional type parameter trait bounds.

Aliased Type

enum LogError<S, SERP, C> {
    StorageError(<S as Storage>::Error),
    SegmentError(SegmentError<<S as Storage>::Error, <SERP as SerializationProvider>::Error>),
    CacheError(<C as Cache<usize, ()>>::Error),
    CacheNotFound,
    BaseIndexLesserThanInitialIndex,
    WriteSegmentLost,
    IndexOutOfBounds,
    IndexGapEncountered,
}

Variants

StorageError(<S as Storage>::Error)

Used to denote errors from the underlying Storage implementation.

SegmentError(SegmentError<<S as Storage>::Error, <SERP as SerializationProvider>::Error>)

Used to denote errors from operations on Segment instances.

CacheError(<C as Cache<usize, ()>>::Error)

Used to denote errors from the SegmentedLog inner cache.

CacheNotFound

Used when the inner cache is not configured while using APIs that expect it.

BaseIndexLesserThanInitialIndex

Used when the resulting base_index of a Segment in the SegmentedLog is lesser than the initial_index configured at the SegmentedLog level.

WriteSegmentLost

Used when the write Segment containing Option is set to None

IndexOutOfBounds

Used when the given index is outside the range [lowest_index, highest_index)

IndexGapEncountered

Used when no Record is found at a valid index inside the range [lowest_index, highest_index]