Trait laminarmq::storage::commit_log::CommitLog

pub trait CommitLog<M, T>: AsyncIndexedRead<Value = Record<M, T>, ReadError = Self::Error> + AsyncTruncate<Mark = Self::Idx, TruncError = Self::Error> + AsyncConsume<ConsumeError = Self::Error> + Sizable {
    type Error: Error;

    // Required method
    fn append<'life0, 'async_trait, X, XBuf, XE>(
        &'life0 mut self,
        record: Record<M, X>
    ) -> Pin<Box<dyn Future<Output = Result<Self::Idx, Self::Error>> + 'async_trait>>
       where X: Stream<Item = Result<XBuf, XE>> + Unpin + 'async_trait,
             XBuf: Deref<Target = [u8]> + 'async_trait,
             XE: 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn remove_expired<'life0, 'async_trait>(
        &'life0 mut self,
        _expiry_duration: Duration
    ) -> Pin<Box<dyn Future<Output = Result<Self::Idx, Self::Error>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

An abstract, append-only, ordered sequence of Record instances.

This trait acts as a generalized storage mechanism for storing our records. All our message-queue server APIs can be expressed using this trait.

Required Associated Types

type Error: Error

Error type associated with CommitLog.

Required Methods

fn append<'life0, 'async_trait, X, XBuf, XE>( &'life0 mut self, record: Record<M, X> ) -> Pin<Box<dyn Future<Output = Result<Self::Idx, Self::Error>> + 'async_trait>>
where X: Stream<Item = Result<XBuf, XE>> + Unpin + 'async_trait, XBuf: Deref<Target = [u8]> + 'async_trait, XE: 'async_trait, Self: 'async_trait, 'life0: 'async_trait,

Appends the given Record at the end of this CommitLog.

The Record may contain a stream of byte slices. Implementations are to exhaustively read the stream and append the corresponding byte slices as a single record.

Returns the index at which the Record was appended.

Provided Methods

fn remove_expired<'life0, 'async_trait>( &'life0 mut self, _expiry_duration: Duration ) -> Pin<Box<dyn Future<Output = Result<Self::Idx, Self::Error>> + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Removes all expired records from this CommitLog.

Returns the number of expired records.

Object Safety

This trait is not object safe.