Trait laminarmq::storage::AsyncIndexedRead

pub trait AsyncIndexedRead {
    type ReadError: Error;
    type Value;
    type Idx: Unsigned + CheckedSub + ToPrimitive + Ord + Copy;

    // Required methods
    fn read<'life0, 'life1, 'async_trait>(
        &'life0 self,
        idx: &'life1 Self::Idx
    ) -> Pin<Box<dyn Future<Output = Result<Self::Value, Self::ReadError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn highest_index(&self) -> Self::Idx;
    fn lowest_index(&self) -> Self::Idx;

    // Provided methods
    fn has_index(&self, idx: &Self::Idx) -> bool { ... }
    fn len(&self) -> Self::Idx { ... }
    fn is_empty(&self) -> bool { ... }
    fn normalize_index(&self, idx: &Self::Idx) -> Option<Self::Idx> { ... }
}

Collection providing asynchronous read access to an indexed set of records (or values).

Required Associated Types

type ReadError: Error

Error that can occur during a read operation.

type Value

Value to be read.

type Idx: Unsigned + CheckedSub + ToPrimitive + Ord + Copy

Type to index with.

Required Methods

fn read<'life0, 'life1, 'async_trait>( &'life0 self, idx: &'life1 Self::Idx ) -> Pin<Box<dyn Future<Output = Result<Self::Value, Self::ReadError>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Reads the value at the given index.

fn highest_index(&self) -> Self::Idx

Index upper exclusive bound

fn lowest_index(&self) -> Self::Idx

Index lower inclusive bound

Provided Methods

fn has_index(&self, idx: &Self::Idx) -> bool

Returns whether the given index is within the index bounds of this collection.

This method checks the following condition:

lowest_index <= idx < highest_index

fn len(&self) -> Self::Idx

Returns the number of values in this collection.

fn is_empty(&self) -> bool

Returns whether this collection is empty.

fn normalize_index(&self, idx: &Self::Idx) -> Option<Self::Idx>

Normalizes the given index between [0, len) by subtracting lowest_index from it.

Returns Some(normalized_index) if the index is within bounds, None otherwise.