Trait laminarmq::storage::AsyncTruncate

pub trait AsyncTruncate {
    type TruncError: Error;
    type Mark: Unsigned;

    // Required method
    fn truncate<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        mark: &'life1 Self::Mark
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::TruncError>> + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Trait representing a truncable collection of records, which can be truncated after a “mark”.

Required Associated Types

type TruncError: Error

Error that can occur during a truncation operation.

type Mark: Unsigned

Type to denote a truncation “mark”, after which the collection will be truncated.

Required Methods

fn truncate<'life0, 'life1, 'async_trait>( &'life0 mut self, mark: &'life1 Self::Mark ) -> Pin<Box<dyn Future<Output = Result<(), Self::TruncError>> + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Truncates this collection after the given mark, such that this collection contains records only upto this “mark”.