Trait laminarmq::storage::Sizable

pub trait Sizable {
    type Size: Unsigned + FromPrimitive + Sum + Ord;

    // Required method
    fn size(&self) -> Self::Size;
}

Tratis representing collections which have a measurable size in number of bytes.

Required Associated Types

type Size: Unsigned + FromPrimitive + Sum + Ord

Type to represent the size of this collection in number of bytes.

Required Methods

fn size(&self) -> Self::Size

Returns the size of this collection in butes.

Object Safety

This trait is not object safe.